Move the config header out of the runtime directory.
