+++ libgalois/include/galois/config.h.in
@@ -1,5 +1,5 @@
-#ifndef GALOIS_RUNTIME_CONFIG_H
-#define GALOIS_RUNTIME_CONFIG_H
+#ifndef GALOIS_CONFIG_H
+#define GALOIS_CONFIG_H
 
 #if !(defined(GALOIS_USE_LONGJMP_ABORT) || defined(GALOIS_USE_EXCEPTION_ABORT))
 #define GALOIS_USE_LONGJMP_ABORT
