+++ libgalois/include/galois/runtime/Context.h
@@ -23,7 +23,7 @@
 #include "galois/MethodFlags.h"
 #include "galois/substrate/PtrLock.h"
 #include "galois/gIO.h"
-#include "galois/runtime/config.h"
+#include "galois/config.h"
 
 #include <boost/utility.hpp>
