+++ libgalois/include/galois/runtime/Executor_ForEach.h
@@ -20,12 +20,12 @@
 #ifndef GALOIS_RUNTIME_EXECUTOR_FOREACH_H
 #define GALOIS_RUNTIME_EXECUTOR_FOREACH_H
 
+#include "galois/config.h"
 #include "galois/gIO.h" // TODO: remove
 #include "galois/Mem.h"
 #include "galois/Timer.h"
 #include "galois/Threads.h"
 #include "galois/Traits.h"
-#include "galois/runtime/config.h"
 #include "galois/runtime/Substrate.h"
 #include "galois/runtime/Context.h"
 #include "galois/runtime/Range.h"
